MommaCat: rip some stuff out and put it in other files
